requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LangChain and LangGraph
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20

# Required dependencies for LangGraph
jsonpatch>=1.33
typing-extensions>=4.5.0

# Utilities
pathlib2>=2.3.7; python_version < '3.4'
